requirements.txt
open3d
pytorch3d
plyfile
opencv-python
lpips
trimesh
tensorboard